+++ R/uploadToDatabaseModelDesign.R
@@ -132,7 +132,7 @@ insertModelDesignSettings <- function(
     conn = conn, 
     resultSchema = databaseSchemaSettings$resultSchema, 
     targetDialect = databaseSchemaSettings$targetDialect,
-    modelType = attr(object$modelSettings$param, 'settings')$modelType,
+    modelType = attr(object$modelSettings$param, 'settings')$modelType, # make this the same as model$trainDetails$modelName?
     json = object$modelSettings, 
     stringAppendToTables = databaseSchemaSettings$stringAppendToResultSchemaTables,
     tempEmulationSchema = databaseSchemaSettings$tempEmulationSchema